feat: add endpoint to generate pdf from html
The endpoint currently supports more than we strictly need. It handles
both file uploads (via multipart/form-data requests) as well as POSTing
html text. Down the line we should add endpoints to generate PDFs based
on a (HTML) file ID, as well as a delta endpoint to make it reactive.
But this will require extra hooks to make it truly useful, so work for
later.